+++ libs/rend/src/rvk/attach.c
@@ -145,8 +145,8 @@ void rvk_attach_pool_destroy(RvkAttachPool* pool) {
   alloc_free_t(g_alloc_heap, pool);
 }
 
-u32 rvk_attach_pool_count(const RvkAttachPool* pool) {
-  return rvk_attach_max_images - (u32)bitset_count(bitset_from_array(pool->emptyMask));
+u16 rvk_attach_pool_count(const RvkAttachPool* pool) {
+  return rvk_attach_max_images - (u16)bitset_count(bitset_from_array(pool->emptyMask));
 }
 
 u64 rvk_attach_pool_memory(const RvkAttachPool* pool) {
